+++ src/lib.rs
@@ -10,3 +10,4 @@ pub(crate) mod colors;
 pub(crate) mod graph;
 pub(crate) mod item;
 pub(crate) mod tree;
+pub(crate) mod utils;